requirements.txt
google-generativeai
python-dotenv
langchain_experimental
langchain_core
google-generativeai==0.3.1
google-ai-generativelanguage==0.4.0
langchain-google-genai
pydantic==1.10.7
ipython
qdrant-client
pillow